burnin_tab additional graphs finalized. formatting adjustments
Corrected the base blue colour to better match the FUS green. added a plt.close('all') to avoid memory leaks if multiple graphs are being plotted. The individual graphs of the pos and negative error are now different colours. Added docstring to moving avg function. Graph with separate error has checkbox as well and sends the boolean value of it's box to the BurninGraphs class call.